Disable our custom swapcontext implementation
Disable it by default, as it is currently broken regardless of the
compiler used.

As it stands, it seems there is not benefit whatsoever to use more
fibers than vCPUs available, so we might want to consider removing this
code completly in the future.

This might be caused by changes in glibc and the Linux kernel, as it
seems they introduced a userland shadow stack. This needs further
investigation.